+++ src/forker.js
@@ -89,16 +89,22 @@ function run(aggregatorFile, options) {
   function step(message) {
     allSteps = allSteps + 1;
 
-    return data => {
+    return (data, options = {}) => {
+      const {message: newMessage, suppress = false} = options;
+
       lastStep = lastStep + 1;
-      process.send({
-        action: "PROGRESS",
-        payload: {
-          message: message,
-          currentStep: lastStep,
-          steps: allSteps
-        }
-      });
+
+      if (!suppress) {
+        process.send({
+          action: "PROGRESS",
+          payload: {
+            message: newMessage || message,
+            currentStep: lastStep,
+            steps: allSteps
+          }
+        });
+      }
+
       return data;
     };
   }